i am stupid

+++ api/src/Features/Services/Delete/Service.cs
@@ -1,5 +1,6 @@
-using Kratos.Api.Common.Types;
 using Microsoft.AspNetCore.Mvc;
+
+using Kratos.Api.Common.Types;
 
 namespace Kratos.Api.Features.Services.Delete;
 
@@ -7,12 +8,19 @@ public class Service([FromServices] IRepository repo)
 {
     public async Task<Result> DeleteAsync(long serviceId, CancellationToken cancellationToken)
     {
-        int rowsAffected = await repo.DeleteAsync(serviceId, cancellationToken);
-        if (rowsAffected == 0)
+        bool exists = await repo.ExistsAsync(serviceId, cancellationToken);
+        if (!exists)
         {
             return Result.NotFoundError("Service not found");
         }
 
+        bool hasChildren = await repo.HasChildrenAsync(serviceId, cancellationToken);
+        if (hasChildren)
+        {
+            return Result.CannotProcessError("Service has children");
+        }
+
+        await repo.DeleteAsync(serviceId, cancellationToken);
         return Result.Success();
     }
 }
